[C-IRIS] Move C-IRIS out of dev folder.
