+++ geometry/optimization/c_iris_collision_geometry.h
