+++ geometry/optimization/c_iris_separating_plane.cc
@@ -1,0 +1,1 @@
+#include "drake/geometry/optimization/c_iris_separating_plane.h"
