--- geometry/optimization/dev/BUILD.bazel
@@ -1,117 +1,0 @@
-# -*- python -*-
-
-load(
-    "@drake//tools/skylark:drake_cc.bzl",
-    "drake_cc_googletest",
-    "drake_cc_library",
-    "drake_cc_package_library",
-)
-load(
-    "@drake//tools/skylark:test_tags.bzl",
-    "mosek_test_tags",
-)
-load("//tools/lint:lint.bzl", "add_lint_tests")
-
-drake_cc_library(
-    name = "c_iris_collision_geometry",
-    srcs = ["c_iris_collision_geometry.cc"],
-    hdrs = ["c_iris_collision_geometry.h"],
-    deps = [
-        "//geometry:geometry_ids",
-        "//geometry:shape_specification",
-        "//geometry/optimization:convex_set",
-        "//multibody/rational:rational_forward_kinematics",
-        "//multibody/tree:multibody_tree_indexes",
-    ],
-)
-
-drake_cc_library(
-    name = "c_iris_separating_plane",
-    srcs = ["c_iris_separating_plane.cc"],
-    hdrs = ["c_iris_separating_plane.h"],
-    deps = [
-        ":c_iris_collision_geometry",
-        "//common/symbolic:polynomial",
-    ],
-)
-
-drake_cc_library(
-    name = "cspace_free_polytope",
-    srcs = ["cspace_free_polytope.cc"],
-    hdrs = ["cspace_free_polytope.h"],
-    deps = [
-        ":c_iris_collision_geometry",
-        ":c_iris_separating_plane",
-        "//common/symbolic:monomial_util",
-        "//solvers:choose_best_solver",
-        "//solvers:mathematical_program",
-        "//solvers:mosek_solver",
-        "//solvers:solve",
-    ],
-)
-
-# test
-
-drake_cc_library(
-    name = "c_iris_test_utilities",
-    testonly = 1,
-    srcs = ["test/c_iris_test_utilities.cc"],
-    hdrs = ["test/c_iris_test_utilities.h"],
-    data = [
-        "test/convex.mtl",
-        "test/convex.obj",
-    ],
-    deps = [
-        ":cspace_free_polytope",
-        "//geometry:scene_graph",
-        "//multibody/plant",
-        "//systems/framework:diagram_builder",
-        "@gtest//:without_main",
-    ],
-)
-
-drake_cc_googletest(
-    name = "c_iris_collision_geometry_test",
-    deps = [
-        ":c_iris_collision_geometry",
-        ":c_iris_test_utilities",
-        "//common/test_utilities:symbolic_test_util",
-    ],
-)
-
-drake_cc_googletest(
-    name = "c_iris_separating_plane_test",
-    deps = [
-        ":c_iris_separating_plane",
-        "//common/test_utilities:symbolic_test_util",
-    ],
-)
-
-drake_cc_googletest(
-    name = "cspace_free_polytope_test",
-    deps = [
-        ":c_iris_test_utilities",
-        ":cspace_free_polytope",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/test_utilities:symbolic_test_util",
-    ],
-)
-
-drake_cc_googletest(
-    name = "cspace_free_polytope_with_mosek_test",
-    timeout = "long",
-    # This test launches 2 threads to test both serial and parallel code paths
-    # in CspaceFreePolytope.
-    tags = mosek_test_tags() + ["cpu:2"] + ["no_memcheck"],
-    use_default_main = False,
-    deps = [
-        ":c_iris_test_utilities",
-        ":cspace_free_polytope",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/test_utilities:symbolic_test_util",
-        "//solvers:mosek_solver",
-        "//solvers:sos_basis_generator",
-    ],
-)
-
-add_lint_tests()
