--- geometry/optimization/dev/c_iris_separating_plane.cc
@@ -1,1 +1,0 @@
-#include "drake/geometry/optimization/dev/c_iris_separating_plane.h"
