+++ geometry/optimization/test/c_iris_separating_plane_test.cc
@@ -1,4 +1,4 @@
-#include "drake/geometry/optimization/dev/c_iris_separating_plane.h"
+#include "drake/geometry/optimization/c_iris_separating_plane.h"
 
 #include <gtest/gtest.h>
