+++ geometry/optimization/test/convex.obj
